Done battle, but must fix returning to overworld scene

+++ Assets/Scripts/Party/HpMpBarManager.cs
@@ -15,11 +15,11 @@ public class HpMpBarManager : MonoBehaviour {
 
 	public void UpdateHp(){
 		hpbar.transform.localScale = 
-			new Vector2(chr.currentHp/chr.calculatedStats.hp, 1);
+			new Vector2((float) chr.currentHp/chr.calculatedStats.hp, 1);
 	}
 
 	public void UpdateMp(){
 		mpbar.transform.localScale = 
-			new Vector2(chr.currentMp/chr.calculatedStats.mp, 1);
+			new Vector2((float) chr.currentMp/chr.calculatedStats.mp, 1);
 	}
 }
